trading-dashboard/backend/requirements.txt
ccxt==4.1.61
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
apscheduler==3.10.4
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0
email-validator==2.2.0
EOF < /dev/null
